fixed mod and div.
git-svn-id: https://pascalinterpreterinjava.googlecode.com/svn/trunk@228 fb5b4802-194b-0410-9e11-5134f6b347ec

+++ src/com/js/interpreter/tokens/OperatorTypes.java
@@ -69,7 +69,7 @@ public enum OperatorTypes {
 		@Override
 		public Object operate(long l1, long l2)
 				throws OperationNotSupportedException {
-			return l1 / l2;
+			return (double) l1 / (double) l2;
 		}
 
 		@Override
